multi-channel changes
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@7479 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/writers/rts2fitsfile.cpp
@@ -315,7 +315,7 @@ int Rts2FitsFile::writeArray (const char *extname, std::list <ColumnData *> & va
 			fits_write_col (ffile, TDOUBLE, i, 1, 1, (*iter)->len, (*iter)->data, &fits_status);
 	}
 
-	for (i = 0; i < values.size (); i++)
+	for (i = 0; i < (int) (values.size ()); i++)
 		delete[] cols[i];
 
 	// move back to primary HDU
